Only perform contextualizeMediaPaths in submission table
